feat(frame): str.get supports Python-style negative indexing
Found via a live-pandas 2.2.3 differential. Series.str.get took a usize index, so
it could not express pandas' Python-style negative indices — str.get(-1) is the
last character. The signature also forced the fp-conformance fixture field
str_get_index to usize, propagating the limitation end to end.

Widen StringAccessor::get to i64 and resolve negatives as i + len, with
out-of-range (either direction) yielding NaN, matching pandas:
  ["hello","ab"].str.get(-1) -> ["o","b"]
  .str.get(-2) -> ["l","a"]
  .str.get(-3) -> ["l", nan]   (in range for "hello", out for "ab")
  .str.get(-10) -> [nan, nan]
fp-conformance str_get_index: Option<usize> -> Option<i64> so fixtures can carry
negative indices (serde-compatible for existing positive values). New test
str_get_negative_index; existing positive/OOB behavior unchanged.

fp-frame lib 2773/0; str_get suite 6/0; fp-conformance compiles; clippy
-p fp-frame code-clean.

+++ crates/fp-conformance/src/lib.rs
@@ -2075,7 +2075,7 @@ pub struct PacketFixture {
     #[serde(default)]
     pub str_slice_end: Option<usize>,
     #[serde(default)]
-    pub str_get_index: Option<usize>,
+    pub str_get_index: Option<i64>,
     #[serde(default)]
     pub str_repeat_n: Option<usize>,
     #[serde(default)]
@@ -3496,7 +3496,7 @@ struct OracleRequest {
     #[serde(default)]
     str_sub: Option<String>,
     #[serde(default)]
-    str_get_index: Option<usize>,
+    str_get_index: Option<i64>,
     #[serde(default)]
     str_fillchar: Option<char>,
     #[serde(default)]
